Add SKIP_WRITE_GIT_CHANGELOG option to setup.py
Updates the write_git_changelog function in setup.py so that
it will skip changelog generation if the SKIP_WRITE_GIT_CHANGELOG
env variable is set.

The motivation for this change is to speed up running sdist which
gets executed at the begining of a tox run. Some projects (like Nova)
can take between 10-20 seconds to generate a changelog in this fashion.

Change-Id: I2e75060d303bc8fae784129e8e42f09922a8b09e