Don't use pbr for pbr.
It turns out, that when you setup_requires pbr from another package, it gets
installed via easy_install instead of via pip. easy_install, craptastic
pile of vomit and excrement that it is, can't manage to process the entry
points in the package it's installing while installing it, so you always
wind up with pbr version=="". Sigh.

Change-Id: Ib2b8e77662286fd4f86e9ccc55680cccc8ce349b