Support lookup of value using "group.key"
Let us check if the opt_name has a '.', if it does then split it
into a group/key and try lookup using that combination. Since
LazyPluggable uses "CONF[self.__pivot]" if we just add this
capability to cfg, we get "LazyPluggable doesn't support
option groups" for free.

Fixes LP #1093043

Change-Id: I9cedcf22014038e9fe4ed5e66ca5427aa99b5091