Explicitly install install_requires.
PBR uses distutils.install instead of setuptools.install to provide
non-egg based installs. This is effectively the same behavior as the
--single-version-externally-managed flag to setuptools, which is what
distro packagers use inside of debs and rpms. We do this so that we
can do things like install manpages and config files.

Unfortunately, this has the unintended consequence of completely
skipping dependency processing.

Add the logic from distribute into our install class to ensure that
the contents of install_requires are installed appropriately.

In the future, we'll want to replace this method with one that uses
pip instead of easy_install. Since we're doing the work now in our
code, we might as well do it well.

Change-Id: Id5465766f762ab37f121ee3d72a82098654e30db