Add an integration test script.
The script is intended to be run on a prepared server such as what
devstack-gate produces.

Test that each of the projects can be installed from scratch without
pbr being previously installed into a plain raw non-distribute-based
virtualenv. This is the closest we can come to testing that pip
installing from the internet is going to work.

We also need to test that this can be done in the various combinations
of setuptools/distribute/pip, and we need to test that pip installing
the repo, the tarball and running python setup.py install work.

Change-Id: If3f17795374cfce726c60de0f09fc2a1e694eea3