Put AUTHORS.in at the top of AUTHORS.
swift has an AUTHORS file that puts a header at the top of the
file with some extra information, then lists the contributors.
If we put AUTHORS.in at the top of the generated file, then we
can put their boilerplate into AUTHORS.in and the resulting file
should look the same.

Change-Id: I2caa4e606dd1199f49c925dccc64d492848ac580