Revert include_package_data change
Setting the value directly in kwargs is not doing what we want it to do.
This is largely because of how the hook system works. When the hooks
process backwards_compat, they do a config.get('backwards_compat',
dict()), which is then written back to the config dict - which means
that blank override values are being fed in.

Later we can go through and just re-engineer how that works. For now,
unbreak nova.

Change-Id: I0c6055253cbc89b6884553e6f2fbfe8a7bbd1953