Stop trying to install old pip and setuptools
Now that pip 1.4 is out, setuptools<0.7 is invalid because 1.4 will not
install 'pre-release' software without being explicitly asked, and all
of the pre 0.7 setuptools releases have a letter in them. Just stop
trying to do it, because we actually don't use it anyway.

Additionally, we were trying to intsall pip 1.3 for a test but were not
installing that into the local mirror. Match them so that they're equal
and can install.

Change-Id: I61b9b810e64b40ceccb5eb0f4b2ddc4faefd8e04