Consume the OpenStack mirror by default
PyPI upstream obviously breaks all the time. Like, literally, I've spent
my whole life dealing with it this past week. In this test, we were
avoiding the OpenStack mirror so that we could be sure to be testing
that we could build the mirror when new requirements are added. However,
the only time that's a concern is on a requirements repo change.

Add an option --no-mirror that will trigger the "don't use OpenStack's
mirrors as an upstream" behavior, and we'll add that option to the
invocation on requirements changes.

Change-Id: I8d01c1fbb8a5df5aa83ffba3b6489e374abef1af