Refactor oneline changelog walking.
We want to suck the changelog into the description as well, but we
don't want all of it there. Making it a generator of formatted
objects lets us apply some logic without conflating things or tying
it up too much. This patch does prep work for the long term goal.

Change-Id: Idd00c219891500210c83faeb0f2d4370add3b069