Support platform-specific requirements files
This patch adds support for choosing a requirements file, based on the
current OS, with the format 'requirements-{osname}'. This can be combined
with the current approach of choosing a requirement file based on the
Python version. The motivation for this is to replace conditional
choosing of packages inside setup.py.

Change-Id: I30c56082bb9b9bcd80128eedcef17b6beb793cf8