Integration test PBR commits
* pbr/packaging.py(_get_version_from_git): Correct a bug which is
causing install failures for PBR-based projects with no Git tags.
This slipped through because integration testing was using the most
recent PBR release rather than the commit being proposed.

* tools/integration.sh: Make sure that if a PBR commit is being
tested then we install and use that source rather than the latest
PBR release.

* tox.ini(testenv:pep8): Get rid of the custom deps list for pep8
and inherit the one from testenv which is already fixed with the
self-dependency. Without this, pep8 jobs fail when run with
Setuptools 8.

Change-Id: Icc7275261e8fc0df8b0ae99184437ca50b89c13a