Fixes Windows generated scripts headers
When installing packages on Windows, generated scripts based on entry
points have wrong Python executable paths in the header. For this
reason, generated executables fail with 'fail to create proccess'.

For example, if the Python executable path is
'C:\Program Files\Python27\python.exe', the scripts will contain
'#!C:ProgramFilesPython27python.exe'

A simple fix is to wrap the Python executable in brackets on Windows.

Change-Id: Id017073c58c162c3e0c546b1ab76a7bea9a9f04e
Closes-Bug: #1416342