Export ChangeLog and AUTHORS in install
readthedocs uses 'setup.py install' to prepare trees for doc creation,
but ChangeLog is not currently created there, and doing so would be
nice. This won't affect develop invocations AFAICT, and even if it
did, the overheads are ~10% of the time to run 0 tests in Nova today -
e.g. quite tolerable.

Change-Id: I7bc18fc9ca2dbe852598cc79b2ad6273fc53557d