Don't attempt to test with 0.6c11 with Py3
Setuptools 0.6c11 is ancient, and doesn't install on Py3, so don't
attempt to do so.  We don't test Py3 in the gate with integration, so
this is a noop there, however this means if we do so in the future
this won't bubble up as a failure.  It also allows me to run this block
of integration tests with Py3 in addtion to Py27 locally.

Change-Id: I1dc6eb7591be847573231ca1056c2cd31b2f673a