Fix pypy soabi tests
Xenial PyPy adds multiarch to the compiled c library, so if PyPy reports
multiarch we add it to the soabi string.

If the installed PyPy is not configured for multiarch, the multiarch
config variable will not exist and `arch` will be None, in which case we
should only update the string if multiarch exists.

Change-Id: Idcd3af4878d5b66c6770f96e7235a7d2f848df81