tox: Don't set skipsdist=True
skipsdist is used to "avoid expensive sdist" but prevents the software
package from being installed in the virtualenv. We currently have this
enabled, but then skip the step by including the current package in the
requirements section, which mitigates the entire thing.

Stop setting skipsdist to True, allowing us to remove '.' from
requirements and use tox the way it's meant to be used.

Change-Id: I543f0c6679c39c7ae438fd1e5fca7175b92ed193