Revert "Remove win32/nt checks for wrapper script gen"
This reverts commit 07de844615eaec520d817558462cda5f91f9b7b8.
The output console_scripts generated after this commit landed are no
longer directly callable from windows machines. Prior to the this
reverts we wrapped console_script entrypoints in a .exe on windows so
you could directly execute them, just like on *nix systems. However,
after the commit we are no longer generating callables on windows.
Instead PBR generates a txt file, without a suffix, that is not a valid
executable on windows. It is exactly the same output as on a *nix
system, but this neglects that scripts with shebangs don't work on
windows. (For example generated output on windows see [1]) The exe files
were needed so that we could directly execute the console scripts on
windows. This commit restores this functionality so that PBR will
properly generate executable console_scripts on windows machines again.

Closes-Bug: #1761134
[1] http://paste.openstack.org/show/722389/

Change-Id: Ifc13879b7f64650d444e3a14df1a53b2172828e4