Update python requires packaging metadata for package
The metadata field, 'python-requires' was introduced in PEP440 and
appears to have superseded 'requires-python'. [1] The metadata that pbr
uses for itself and it's test package was using the old incorrect field,
this commit updates it to be the proper metadata to show a working
example. To a certain extent it probably doesn't matter because it's very
unlikely someone will try to install pbr with python 2.5, but it's more
about making sure that pbr itself is leading by example and that it's own
config actually works.

[1] https://github.com/mtreinish/stestr/issues/292

Change-Id: I5eb14d490a3731b6b95d7a1137c33ec2d73f193f