More easy_install.ScriptWriter.get_header()
Setuptools 12.0 deprecated easy_install.get_script_header() when it
was released five years ago, and if called with PYTHONWARNINGS=error
raises a deprecation exception. Switch to the preferred class
method, but keep the original call as a fallback since it won't work
on older Setuptools versions (such as the 3.3 shipped with Ubuntu
Trusty). This continues 9a219f9a4f47243123d50d30938506bcc792e55d to
remove a second call for the same deprecated function.

Change-Id: I365075c2cff88283f92298ce2f5e53d263dc7f4b