Run tempest-full for stable/train
The tempest-full job runs Python 2.7 (as opposed to
tempest-full-py3) so will want to install Python library packages
which have been dropped in Focal in favor of their python3-.*
replacements. Switch this job to use stable/train branches of
required projects as they still expected 2.7 to work, and as a side
effect this will cause the appropriate nodeset to be chosen
automatically.

Change-Id: Id95eaae9351e8664aa1fa48a7b4bf59e4c463834