Revert "Map requires-python to python-requires"
This reverts commit d5da2121f875109a94809eeb8aa271fd5a0b3edd. This was
done incorrectly. The 'D1_D2_SETUP_ARGS' dictionary modified here is a
mapping of distutils (and by extension, setuptools) supported options to
their legacy distutils2 equivalents, and is used to transform the latter
to the former. For example, consider the following entry:

  "long_description": ("metadata", "description"),

This will transform an entry like so in a 'setup.cfg' file:

  [metadata]
  description = ...

Into the following equivalent call to setuptools' 'setup' function:

  setup(
      ...,
      long_description='...',
      ...
  )

By that measure, the entry added in this change, namely:

  "requires_python": ("metadata", "python_requires"),

Will result in a 'requires_python' argument being added to the 'setup'
call. That's not what we want. Rather, we likely want to transform a
'[metadata] requires_python' entry to a 'python_requires' argument to
'setup'.

The change is reverted here in expectation of a forthcoming proper fix.

Change-Id: I7158e75934535b80eb60d6b4b7dcef08866b5bab
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>