Update dependencies to avoid failure with old pip
We do perform linting via pep8 job which runs on modern python only.
This fixes an error that recently broke py27 job due to failure to
install hacking, where hacking is in fact not even needed by it.

We also need to add some extra conditions for mock to avoid possible
failure with older pip version.

Change-Id: I5007154caa7493652a424a4fad4918b7dbefcfb9