Run python3.9 test jobs
We want to run python3.9 test jobs for PBR too. Also we explicitly list
out all the python jobs because the various openstack project-templates
for this are release specific and PBR is not release specific. PBR must
run against all of the releases at the same time. Just spell it out then
it becomes easy to add new versions of python.

Change-Id: I9c85731040a73fa57e45ad8da3924c4d13c5affc