PBR package testing improvements
We simplify virtualenv setup for PBR package installation testing by
switching to a simple file path for the PBR installation rather than a
full url with egg specifier. The egg specifer isn't necessary and
neither is the url and simpler is easier to understand.

We also add the verbose flag to the pip installations of tools into that
virtualenv to better understand problems when they occur.

Finally we add the option to skip writing package files if their
contents are empty. This allows us to ensure that we can omit certain
files which will be useful for the next change.

Change-Id: I1f081d29d62c905c1bb1abee5402b0990c83e948