Fix integration testing of current change
The detection of whether or not we are running against a zuul change was
completely broken in tools/integration.sh. It was attempting to check
against zuulv2 behaviors which haven't worked for years.

Fix this by comparing the local master branch against the origin/master
branch as zuul manages these refs to point master at the current chagne
to master and origin/master at a parent. If these two things differ then
we are running against a zuul change (and not for example in periodic).

This is still potentially incomplete as we may want periodic jobs to
install from master rather than using the latest release.

Change-Id: Ia7aea1750ab2ec4bf0eace27744720f93fecceaa