Use pyproject.toml
Enough time has passed with other systems using pbr with pyproject.toml
build specifications to be confident that this should work for PBR
itself at this point. Make the switch.

Part of what is motivating this change is a desire to make use of PBR
more consistent with modern python package tooling expectations. For
example setuptools is no longer included by default in python3.12 and
newer. Through the use of pyproject.toml we can capture that setuptools
is a dependency for PBR installation.

Python3.6 presents a problem because it supports a version of pip
(21.3.1) with pep 660 support but there is no setuptools version with
python3.6 and pep660 support. Unfortunately, when you have a
pyproject.toml in your project pip uses this as an indicator to
unconditionally perform pep660 operations and use editable wheel
installs for development installs when requested. This breaks with
python3.6 due to the lack of support in compatible setuptools when we
add the pyproject.toml to PBR. To work around this in testing we simply
tell tox to not use a development install of PBR under python3.6.

This has the potential to break PBR with python3.6 (and older too if the
pip they run under has pep660 support) in the wild when installing
editable packages if projects have pyproject.toml files. This seems
unlikely as project versions supporting python3.6 are unlikely to have
pyproject.toml files (due to the existing incompatibilities). The
workarounds are for people to not use editable installs or to use an
older pip version.

Change-Id: I66dfd85972260e9f36b7bc35d26483d04af3690c