Set up PBR with setuptools as a dependency
For python3.12 and newer setuptools is no longer present by default.
Express this runtime dependencies for PBR via a normal
requirements.txt. The expectation is that anyone installing a package
with PBR as the build tool will use a pyproject.toml file and if using
PBR at runtime will list PBR as a runtime dependency which should pull
in setuptools for the runtime dep.

We modify the integration tests to expand the coverage of our test
project installation as a result. In particular we want to ensure this
forward looking set of behaviors works as expected while still checking
existing behaviors continue to work for packages that haven't updated
yet.

Depends-On: https://review.opendev.org/c/openstack/pbr/+/938076
Change-Id: I7ef7472ae48c65a47581694f2f1fcc4edbb1d6ad